imageCompaction/requirements.txt
Pillow>=9.4.0
requests>=2.31.0
boto3